removed unnecessary getDirectFlags method and some minor cleanup

+++ core/src/main/java/com/graphhopper/routing/VirtualEdgeIterator.java
@@ -281,7 +281,7 @@ class VirtualEdgeIterator implements EdgeIterator, CHEdgeIteratorState {
     }
 
     @Override
-    public int getMergeStatus(long flags) {
+    public int getMergeStatus(int flags) {
         throw new UnsupportedOperationException("Not supported.");
     }
 }
