+++ core/src/main/java/com/graphhopper/routing/VirtualEdgeIteratorState.java
@@ -255,7 +255,7 @@ public class VirtualEdgeIteratorState implements EdgeIteratorState, CHEdgeIterat
     }
 
     @Override
-    public int getMergeStatus(long flags) {
+    public int getMergeStatus(int flags) {
         throw new UnsupportedOperationException("Not supported.");
     }
